requirements.txt
absl-py==1.2.0
aiosignal==1.3.1
alabaster==0.7.12
altair==5.0.1
anyio==3.6.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asciitree==0.3.3
asttokens==2.0.8
astunparse==1.6.3
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blessed==1.19.1
blinker==1.6.2
build==0.8.0
cachetools==5.2.0
cachey==0.2.1
cellpose @ git+https://github.com/kevinjohncutler/cellpose@54938e2391b5c286f4e5712115eea3092f505653
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.1.0
colour==0.1.5
commonmark==0.9.1
concave-hull @ file:///home/ameyasu/cuda_ws/src/concave-hull
cupy-cuda117==10.6.0
cycler==0.11.0
Cython==0.29.33
darkdetect==0.7.1
dask==2022.9.0
debugpy==1.6.3
decorator==4.4.2
defusedxml==0.7.1
delta2==2.0.5
distlib==0.3.6
docstring-parser==0.14.1
docutils==0.19
edt==2.3.0
elasticdeform @ file:///home/ameyasu/cuda_ws/elasticdeform
entrypoints==0.4
executing==1.0.0
fasteners==0.18
fastjsonschema==2.16.1
fastremap==1.13.3
fastrlock==0.8
ffmpeg-python==0.2.0
filelock==3.9.0
flatbuffers==1.12
flyingcircus==0.1.4.0
flyingcircus-numeric==0.1.1.2
freetype-py==2.3.0
frozenlist==1.3.3
fsspec==2022.8.2
future==0.18.2
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.10.2
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.48.1
h5py==3.7.0
HeapDict==1.0.1
hsluv==5.0.3
hyperopt==0.2.7
idna==3.3
imagecodecs==2022.9.26
imageio==2.21.2
imageio-ffmpeg==0.4.7
imagesize==1.4.1
importlib-metadata==6.8.0
ipykernel==6.15.2
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.10
jsonschema==4.15.0
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyter_client==7.3.5
jupyterlab==3.4.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jupyterlab_server==2.15.1
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
llvmlite==0.39.1
locket==1.0.0
lxml==4.9.1
magicgui==0.5.1
mahotas @ git+https://github.com/luispedro/mahotas@ba045e0dd9454662319c367cdc6535f87f826beb
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.4.3
matplotlib-inline==0.1.6
matplotlib-scalebar==0.7.2
mgen @ git+https://github.com/kevinjohncutler/mgen@011509a3b184a4efa97676932a502c636278ae5a
mistune==2.0.4
moviepy==1.0.3
msgpack==1.0.4
napari==0.4.15
napari-console==0.0.4
napari-plugin-engine==0.2.0
napari-svg==0.1.6
natsort==7.1.1
nbclassic==0.4.3
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
ncolor @ git+https://github.com/kevinjohncutler/ncolor@e706b53db7c03595207aa8d15411ba2aba60056c
nest-asyncio==1.5.5
networkx==2.8.6
notebook==6.4.12
notebook-shim==0.1.0
npe2==0.6.1
numba==0.56.4
numcodecs==0.10.2
numpy==1.23.4
numpydoc==1.4.0
nvidia-cudnn==8.2.0.51
nvidia-pyindex==1.0.9
oauthlib==3.2.0
omnipose @ git+https://github.com/kevinjohncutler/omnipose.git@85455a6e754a4c368b1e3a8e877d16fcfcc6981c
opencv-python==4.5.5.64
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.3
partd==1.3.0
pbr==5.10.0
pep517==0.13.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
Pint==0.19.2
platformdirs==3.0.0
pooch==1.6.0
proglog==0.1.10
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.20.3
psfmodels==0.3.2
psutil==5.9.2
psygnal==0.3.5
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==12.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.3
pycocotools @ git+https://github.com/philferriere/cocoapi.git@2929bd2ef6b451054755dfd7ceb09278f935f7ad#subdirectory=PythonAPI
pycparser==2.21
pydantic==1.10.1
pydeck==0.8.1b0
pyglet==1.5.21
Pygments==2.13.0
Pympler==1.0.1
pymunk==6.2.0
PyOpenGL==3.1.6
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyQt6==6.4.0
PyQt6-Qt6==6.4.0
PyQt6-sip==13.4.0
pyqtdarktheme @ git+https://github.com/kevinjohncutler/omnipose-theme@873db702049049813666e1b6a369fb9deda6eb28
pyqtgraph==0.13.1
pyrsistent==0.18.1
python-dateutil==2.8.2
pytomlpp==1.0.11
pytorch-ranger==0.1.1
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.2.1
qtconsole==5.3.2
QtPy==2.2.0
raster-geometry==0.1.4.1
ray==2.3.0
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.5.1
rsa==4.9
scikit-image==0.18.3
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.2
Send2Trash==1.8.0
setuptools-scm==7.0.5
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==5.1.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.5.0
stqdm==0.0.5
streamlit==1.24.1
superqt==0.3.5
SyMBac @ git+https://github.com/georgeoshardo/SyMBac@cd54061ebfe3678834faf1577da571d8fb8475c2
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
tensorflow==2.8.0
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.15.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tifffile==2021.10.12
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.12.1+cu116
torch-optimizer==0.3.0
torchaudio==0.12.1+cu116
torchvision==0.13.1+cu116
tornado==6.2
tqdm==4.62.3
traitlets==5.3.0
typer==0.6.1
typing_extensions==4.3.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.12
validators==0.20.0
virtualenv==20.19.0
vispy==0.11.0
watchdog==3.0.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.14.1
zarr==2.13.3
zipp==3.16.2
